requirements.txt
Flask==3.0.0
pandas==2.1.4
folium==0.15.1
matplotlib==3.8.2
seaborn==0.13.0
fpdf2==2.7.6
Jinja2==3.1.2
Werkzeug==3.0.1
openpyxl==3.1.2
python-dateutil==2.8.2
